Compatibility patch for other mods that add rubies and sapphires.
Thanks to HybridDog for pointing this out. This update renames the mod
folders for my ruby and sapphire mods so that they don't get mixed up
with anyone else's, like Zeg9's Weird Ores mod (https://github.com/Zeg9
/minetest-weirdores).

Also added some aliases for that mod so that their rubies are replaced
with mine. This seems to work in Worldgen but I haven't seen any rubies
that have already been mined be replaced, so keep an eye out for that.